+++ src/input/exports.ts
@@ -1,9 +1,35 @@
 import * as ts from 'typescript';
+import { getDeclarationFromNode, getParameterName } from '../helpers';
 import { DeclVisitorContext } from '../types';
 
 export function includeExports(context: DeclVisitorContext, key: string, symbol: ts.Symbol) {
-  const exports = context.checker.getExportsOfModule(symbol);
-  context.availableImports[key] = exports
-    .map(exp => exp.valueDeclaration || exp.declarations?.[0])
-    .filter(m => m !== undefined);
+  const defs = {};
+
+  context.checker.getExportsOfModule(symbol).forEach(exp => {
+    const decl = exp.valueDeclaration || exp.declarations?.[0];
+
+    if (!decl) {
+      // skip - not really defined
+    } else if (ts.isExportSpecifier(decl)) {
+      defs[decl.name.text] = getDeclarationFromNode(context.checker, decl);
+    } else if (ts.isExportAssignment(decl)) {
+      defs['default'] = getDeclarationFromNode(context.checker, decl);
+    } else if (ts.isVariableDeclaration(decl)) {
+      defs[getParameterName(decl.name)] = decl;
+    } else if (
+      ts.isFunctionDeclaration(decl) ||
+      ts.isInterfaceDeclaration(decl) ||
+      ts.isClassDeclaration(decl) ||
+      ts.isTypeAliasDeclaration(decl) ||
+      ts.isEnumDeclaration(decl)
+    ) {
+      defs[decl.name.text] = decl;
+    } else if (ts.isMethodDeclaration(decl) || ts.isPropertyDeclaration(decl) || ts.isModuleDeclaration(decl)) {
+      // skip - mostly from ambient modules
+    } else {
+      context.warn(`Skipping import of unknown node (kind: ${decl.kind}).`);
+    }
+  });
+
+  context.availableImports[key] = defs;
 }