+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog for dets
 
+## 0.5.2
+
+- Fixed referencing of indirect symbols
+- Improved usage of default imports
+
 ## 0.5.1
 
-- Fixed undefined typeof
+- Fixed undefined `typeof` declaration
 
 ## 0.5.0